REFACTOR: created view struct to manage displaying the testing view

+++ src/constants.h
@@ -1,7 +1,7 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-#define WORDS_PER_TEST 50
+#define WORDS_PER_TEST (size_t)50
 #define MAX_CHARS_PER_LINE 66
 #define MAX_TEST_WIN_ROWS 3
